refactor: Member of Infos class are moved to info.hpp
